update to use stactools package
